Refactored factories to split model ones vs. reddit ones
